--HG--
branch : perf

+++ src/Tools/Orchard/Program.cs
@@ -3,7 +3,7 @@
 namespace Orchard {
     public class Program {
         public static int Main(string[] args) {
-            return new OrchardHost(Console.In, Console.Out, args).Run();
+            return (int)new OrchardHost(Console.In, Console.Out, args).Run();
         }
     }
 }
